requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0
Flask-Mail==0.9.1